Refactor the `DecisionScreen`
