+++ lib/app/view/app.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:habit_quest/features/decision/presentation/decision_screen.dart';
+import 'package:habit_quest/features/decision/presentation/views/decision_screen.dart';
 import 'package:habit_quest/features/theme/presentation/bloc/theme_bloc.dart';
 import 'package:habit_quest/features/theme/presentation/bloc/theme_state.dart';
 import 'package:habit_quest/l10n/l10n.dart';
